switch to using dotnet build
using dotnet publish changes how the output folder is structured, so plugins wont be built in the correct directory and the built artifact will not have any plugins loaded. a side effect of switching to dotnet build is not being able to run an action to create a linux build unfortunately.